# Conflicts:
#	src/main/java/com/example/sideproject/domain/chat/service/ChatService.java

+++ src/main/java/com/example/sideproject/domain/applicant/service/ApplicantService.java
@@ -7,6 +7,7 @@ import com.example.sideproject.domain.applicant.entity.Applicant;
 import com.example.sideproject.domain.applicant.entity.ApplicationStatus;
 import com.example.sideproject.domain.applicant.repository.ApplicantRepository;
 import com.example.sideproject.domain.applicant.repository.query.ApplicantQueryRepository;
+import com.example.sideproject.domain.notification.service.ApplicantNotificationService;
 import com.example.sideproject.domain.project.entity.Project;
 import com.example.sideproject.domain.project.service.ProjectService;
 import com.example.sideproject.domain.user.entity.User;
@@ -24,6 +25,7 @@ public class ApplicantService {
     private final ApplicantRepository applicantRepository;
     private final ProjectService projectService;
     private final ApplicantQueryRepository applicantQueryRepository;
+    private final ApplicantNotificationService applicantNotificationService;
 
     /**
      * 프로젝트 지원
@@ -44,6 +46,13 @@ public class ApplicantService {
                 .status(ApplicationStatus.unviewed)
                 .build();
 
+        applicantNotificationService.registerApplicant(
+                projectId,
+                project.getTitle(),
+                applicant.getPosition(),
+                project.getUser().getId()
+        );
+
         return applicantRepository.save(applicant).getId();
     }
 
@@ -53,9 +62,23 @@ public class ApplicantService {
     @Transactional
     public void updateStatus(User user, Long projectId, Long applicantId, ApplicationStatus status) {
         Project project = projectService.findProject(projectId);
-        Applicant applicant = applicantRepository.findByIdAndProjectAndUser(applicantId, project, user)
+        if (!project.isProjectLeader(user.getId())) {
+            throw new CustomException(ErrorType.APPLICANT_NOT_FOUND);
+        }
+
+        Applicant applicant = applicantRepository.findByIdAndProject(applicantId, project)
                 .orElseThrow(() -> new CustomException(ErrorType.APPLICANT_NOT_FOUND));
+
         applicant.updateStatus(status);
+
+        if (status.isNotify()) {
+            applicantNotificationService.changeApplicantStatus(
+                    projectId,
+                    project.getTitle(),
+                    status,
+                    applicant.getUser().getId()
+            );
+        }
     }
 
     /**
@@ -63,8 +86,13 @@ public class ApplicantService {
      */
     public void cancel(User user, Long projectId, Long applicantId) {
         Project project = projectService.findProject(projectId);
-        Applicant applicant = applicantRepository.findByIdAndProjectAndUser(applicantId, project, user)
+        Applicant applicant = applicantRepository.findByIdAndProject(applicantId, project)
                 .orElseThrow(() -> new CustomException(ErrorType.APPLICANT_NOT_FOUND));
+
+        if (!applicant.isOwn(user.getId())) {
+            throw new CustomException(ErrorType.APPLICANT_NOT_FOUND);
+        }
+
         applicantRepository.delete(applicant);
     }
